Fix cohort
