+++ src/Cohort.ts
@@ -31,6 +31,17 @@ import {
   dataDBCohortWithNumFilter,
   getCohortData,
   getCohortSize,
+  sizeDBCohortDepletionScoreFilter,
+  sizeDBCohortGeneWithEqualsFilter,
+  sizeDBCohortGeneWithNumFilter,
+  sizeDBCohortPanelAnnotationFilter,
+  sizeDBCohortWithEqualsFilter,
+  sizeDBCohortWithNumFilter,
+  updateCohortName,
+} from './base/rest';
+import type { Task } from './Tasks';
+import { deepCopy, handleDataLoadError, handleDataSaveError, log, mergeTwoAllFilters } from './util';
+import {
   ICohortDBDataParams,
   ICohortDBParams,
   ICohortDBSizeParams,
@@ -50,17 +61,8 @@ import {
   ICohortRow,
   IEqualsList,
   INumRange,
-  sizeDBCohortDepletionScoreFilter,
-  sizeDBCohortGeneWithEqualsFilter,
-  sizeDBCohortGeneWithNumFilter,
-  sizeDBCohortPanelAnnotationFilter,
-  sizeDBCohortWithEqualsFilter,
-  sizeDBCohortWithNumFilter,
-  updateCohortName,
   valueListDelimiter,
-} from './base/rest';
-import { mergeTwoAllFilters, Task } from './Tasks';
-import { deepCopy, handleDataLoadError, handleDataSaveError, log } from './util';
+} from './base/interfaces';
 
 type ICreateMethod<T> = (
   params:
