+++ src/app/CoralApp.ts
@@ -24,8 +24,8 @@ import { setDatasetAction } from '../Provenance/General';
 import { getDBCohortData } from '../base/rest';
 import { Task } from '../Tasks';
 import Taskview from '../Taskview/Taskview';
-import deleteModal from './templates/DeleteModal.html';
-import welcomeHtml from './templates/Welcome.html';
+import deleteModal from '../templates/DeleteModal.html';
+import welcomeHtml from '../templates/Welcome.html';
 import { getAnimatedLoadingText, handleDataLoadError, log } from '../util';
 import { CohortSelectionEvent, ConfirmTaskEvent, CONFIRM_TASK_EVENT_TYPE, PreviewConfirmEvent } from '../base/events';
 import { idCellline, idCovid19, idStudent, idTissue, IEntitySourceConfig } from '../config/entities';
