+++ src/base/index.ts
@@ -1,0 +1,2 @@
+export * from './rest';
+export * from './interfaces';
