+++ src/data/Attribute.ts
@@ -1,15 +1,15 @@
-import { IAllFilters, IServerColumn } from 'tdp_core';
-import { IDataSubtypeConfig, IDataTypeConfig, resolveDataTypes } from 'tdp_publicdb';
-import {
-  Cohort,
-  createCohortWithDepletionScoreFilter,
-  createCohortWithEqualsFilter,
-  createCohortWithGeneEqualsFilter,
-  createCohortWithGeneNumFilter,
-  createCohortWithNumFilter,
-  createCohortWithPanelAnnotationFilter,
-} from '../Cohort';
+import { IAllFilters } from 'tdp_core';
+import { IDataTypeConfig, IDataSubtypeConfig, resolveDataTypes } from 'tdp_publicdb';
+import { IServerColumn } from 'visyn_core';
 import { ICohort } from '../app/interfaces';
+import {
+  HistRouteType,
+  ICohortDBDepletionScoreParams,
+  ICohortDBGeneScoreParams,
+  ICohortDBPanelAnnotationParams,
+  IEqualsList,
+  INumRange,
+} from '../base/interfaces';
 import {
   getCohortData,
   getCohortDepletionScore,
@@ -17,108 +17,23 @@ import {
   getCohortHist,
   getCohortPanelAnnotation,
   getCohortSize,
-  HistRouteType,
-  ICohortDBDepletionScoreParams,
-  ICohortDBGeneScoreParams,
   ICohortDBHistDataParms,
   ICohortDBHistPanelParms,
   ICohortDBHistScoreDepletionParms,
   ICohortDBHistScoreParms,
-  ICohortDBPanelAnnotationParams,
-  IEqualsList,
-  INumRange,
 } from '../base/rest';
-import { IOption, IScoreOption, IServerColumnOption, ISpecialOption, OptionType } from '../Taskview/SearchBar';
-import { deepCopy, getSessionStorageItem, IAttributeFilter, log, setSessionStorageItem } from '../util';
+import {
+  createCohortWithDepletionScoreFilter,
+  createCohortWithEqualsFilter,
+  createCohortWithGeneEqualsFilter,
+  createCohortWithGeneNumFilter,
+  createCohortWithNumFilter,
+  createCohortWithPanelAnnotationFilter,
+} from '../Cohort';
+import { deepCopy, getSessionStorageItem, IAttributeFilter, setSessionStorageItem } from '../util';
 import { easyLabelFromFilter, easyLabelFromFilterArray, niceName } from '../utils/labels';
-import { ISpecialAttribute } from './SpecialAttribute';
-
-export type AttributeType = 'categorical' | 'number' | 'string';
-
-export type ScoreType = 'depletion' | 'expression' | 'copy_number' | 'mutation';
-
-export type IdValuePair = {
-  id: string;
-  [key: string]: any;
-};
-
-export interface IAttributeJSON {
-  option: IOption;
-  currentDB: string;
-  currentView: string;
-}
-
-/**
- * base type for ServerColumns and ScoreColumn
- * id, view, and database are needed for the methods in rest.ts
- */
-export interface IAttribute {
-  /**
-   * id database access the data
-   */
-  id: string;
-
-  /**
-   * datakey with which the attributes data can be accessed frontend
-   * necessary as GeneScoreAttributes have the same id for TPM, Copy Number, etc of the same gene
-   */
-  dataKey: string;
-
-  /**
-   * view to access the data
-   */
-  view: string;
-
-  /**
-   * database to access the data
-   */
-  database: string;
-
-  /**
-   * label of this column for display to the user by default the column name
-   */
-  label: string;
-
-  /**
-   * column type
-   */
-  type: AttributeType;
-
-  /**
-   * the categories in case of type=categorical
-   */
-  categories?: string[];
-
-  /**
-   * the minimal value in case of type=number
-   */
-  min?: number;
-
-  /**
-   * the maxmial value in case of type=number
-   */
-  max?: number;
-
-  /**
-   * wether the attribute is best shown on a log scale
-   */
-  preferLog: boolean;
-
-  getData(cohortDbId: number, filters?: IAllFilters): Promise<IdValuePair[]>; // IRow has _id but IScoreRow has not
-
-  getHist(dbId: number, filters?: IAllFilters, bins?: number): Promise<{ bin: string; count: number }[]>;
-
-  getHistWithStorage(
-    histType: HistRouteType,
-    params: ICohortDBHistDataParms | ICohortDBHistScoreParms | ICohortDBHistScoreDepletionParms | ICohortDBHistPanelParms,
-  ): Promise<{ bin: string; count: number }[]>;
-
-  getCount(cohortDbId: number, filters?: IAllFilters): Promise<number>;
-
-  filter(cht: ICohort, filter: INumRange[] | IEqualsList): Promise<ICohort>;
-
-  toJSON();
-}
+import { AttributeType, IAttribute, IAttributeJSON, IdValuePair, OptionType, ScoreType } from './IAttributue';
+import { ISpecialAttribute } from './ISpecialAttribute';
 
 export abstract class Attribute implements IAttribute {
   public preferLog = false;
@@ -532,40 +447,6 @@ export class PanelScoreAttribute extends AScoreAttribute {
       currentDB: this.database,
       currentView: this.view,
     };
-  }
-}
-
-export function toAttribute(option: IOption, currentDB, currentView): IAttribute {
-  if (option.optionType === 'dbc') {
-    if (option.optionData && (option as ISpecialOption).optionData.spAttribute) {
-      // Create Special Attribtues
-      // if (option.optionData.spA === 'treatment') {
-      log.debug('create special Attribute: ', option.optionId);
-      log.debug('special Attribute object: ', option.optionData.spAttribute);
-      return new SpecialAttribute(
-        option.optionId,
-        currentView,
-        currentDB,
-        (option as ISpecialOption).optionData.spAttribute,
-        (option as ISpecialOption).optionData.attrOption,
-      );
-    }
-    // Create Attribute
-    return new ServerColumnAttribute(option.optionId, currentView, currentDB, (option as IServerColumnOption).optionData.serverColumn);
-  }
-  // Create ScoreAttribute
-  if (option.optionType === 'gene') {
-    return new GeneScoreAttribute(
-      option.optionId,
-      option.optionText,
-      currentView,
-      currentDB,
-      (option as IScoreOption).optionData.type,
-      (option as IScoreOption).optionData.subType,
-    );
-  }
-  if (option.optionType === 'panel') {
-    return new PanelScoreAttribute(option.optionId, currentView, currentDB, 'categorical');
   }
 }
