+++ src/data/SpecialAttribute.ts
@@ -1,62 +1,14 @@
 import _ from 'lodash';
 import { IAllFilters } from 'tdp_core';
-import { Cohort, createCohortWithTreatmentFilter } from '../Cohort';
-import { getRootCohort } from '../cohortview';
-import {
-  getCohortData,
-  HistRouteType,
-  ICohortDBHistDataParms,
-  ICohortDBHistPanelParms,
-  ICohortDBHistScoreDepletionParms,
-  ICohortDBHistScoreParms,
-  IEqualsList,
-  INumRange,
-} from '../base/rest';
+import { getCohortData, ICohortDBHistDataParms, ICohortDBHistPanelParms, ICohortDBHistScoreDepletionParms, ICohortDBHistScoreParms } from '../base/rest';
 import { deepCopy, getSessionStorageItem, log, setSessionStorageItem } from '../util';
 import { niceName } from '../utils/labels';
-import { AttributeType, IdValuePair } from './Attribute';
 import { ICohort } from '../app/interfaces';
-
-export interface ISpecialAttribute {
-  readonly overrideSearchBarDetails: boolean;
-  readonly overrideGetData: boolean;
-  readonly overrideGetHist: boolean;
-  readonly overrideGetCount: boolean;
-  readonly overrideFilter: boolean;
-
-  /**
-   * id database access the data
-   */
-  readonly id: string;
-
-  label: string;
-  type: AttributeType;
-  dataKey: string;
-  /**
-   * Possible options the attribute could be formated
-   */
-  options: { id: string; name: string }[];
-
-  /**
-   * Option that is should be used to format the data
-   */
-  attributeOption: string;
-
-  getDetailForSearchBar(): HTMLDivElement;
-
-  getData(cohortDbId: number, filters?: IAllFilters): Promise<IdValuePair[]>; // IRow has _id but IScoreRow has not
-
-  getHist(dbId: number, filters?: IAllFilters, bins?: number): Promise<{ bin: string; count: number }[]>;
-
-  getHistWithStorage(
-    histType: HistRouteType,
-    params: ICohortDBHistDataParms | ICohortDBHistScoreParms | ICohortDBHistScoreDepletionParms | ICohortDBHistPanelParms,
-  ): Promise<{ bin: string; count: number }[]>;
-
-  getCount(cohortDbId: number, filters?: IAllFilters): Promise<number>;
-
-  filter(cht: ICohort, filter: INumRange[] | IEqualsList, label: string): Promise<ICohort>;
-}
+import { createCohortWithTreatmentFilter } from '../Cohort';
+import { getRootCohort } from '../cohortview';
+import { INumRange, IEqualsList, HistRouteType } from '../base';
+import { AttributeType, IdValuePair } from './IAttributue';
+import { ISpecialAttribute } from './ISpecialAttribute';
 
 export class SATreatment implements ISpecialAttribute {
   readonly overrideSearchBarDetails: boolean = true;
