+++ src/index.ts
@@ -3,11 +3,11 @@ export * from './data';
 export * from './Overview';
 export * from './Taskview';
 
-export * from './app/Coral';
+export * from './app';
 export * from './Cohort';
 export * from './CohortRepresentations';
 export * from './cohortview';
-export * from './base/rest';
+export * from './base';
 
 export * from './util';
 export * from './utils';
